Compile clean under GCC 15 and hold Linux to -Werror too
A clean Debug build of d1cd5c4 on Linux/GCC 15.2 reported 849 warnings
from five causes, every one of them in our own sources -- so SW_WERROR
now defaults ON there as it does on Apple, and the fixes are what the
warnings were asking for rather than suppressions.

Restrict and const qualifiers on return types, 813 of the 849. A
top level qualifier on a return type is dropped by the language, so
`T *LE_RESTRICT` as Span's and AlignedBuffer's iterator alias promised
nothing everywhere begin() and end() were called -- and it was a promise
a span cannot keep anyway, since begin() and end() of one, or any two
copies of one, are pointers into the same array. The aliases are plain
pointers now and the one declaration in the tree that was really getting
a restrict pointer out of one (processor.cpp's WOLA walk) spells the
qualifier itself. Same shape in ParameterLabelGetter and
ModuleConstructor's result_type, and in MemoryMapping::operator[],
which returned `char const` by value.

LE_ASSUME( &reference ), eight warnings across three files. It told the
optimiser that the address of a reference is not null, which the
language already guarantees; GCC reports it as a nonnull argument
compared to NULL. The dynamic_cast assertions beside the two in
moduleControl.cpp check something real and stay. The one in
moduleGUI.hpp did not warn only because nothing calls that inline
function yet; it goes with them.

AutomatedModuleChain::noModule was an unnamed enum, so
`slotIsFull ? module.effectTypeIndex() : noModule` mixed an enumerated
and a non-enumerated type in both places that ask which effect is in a
slot. It is a std::int8_t constant now -- the module chain parameter's
own value type, which is what those conditionals produce and what every
caller already holds.

The FFT's two bin count assertions compared a std::size_t against an
int, which is what a `#pragma warning( disable : 4389 )` had been
bracketing those two functions for on MSVC; saying which type the count
is in removes both the mismatch and the pragma. And Knob's film strip
paint() hid juce::Slider::paint( Graphics & ) from every derived knob --
a `#pragma clang diagnostic ignored "-Woverloaded-virtual"` on one
compiler, 16 warnings on the other. It is paintFilmStrip() now; both
call sites already had to qualify it to say which paint they meant.

Turning the default on for Linux then meant Release had to be clean as
well, and -O3 sees what -O0 cannot:

  - Seven `reinterpret_cast<int const &>( aFloat )`s in math.cpp, plus
    the rest of the file's float-as-integer reads that GCC has not
    got to yet. That is not a reinterpretation of bits but a read of a
    float through an int lvalue, so the compiler may assume it never
    happens and reorder the read against the write that produced the
    value. std::bit_cast is the C++20 spelling of what all of them
    meant, defined, and the same instruction.

  - The module chain's post-increment advanced and then returned
    `reinterpret_cast<chain_iterator &>( node->previous_ )` -- the same
    violation, and a writable alias of a live list link handed to the
    caller. It reached the right node only because the node it had moved
    to points back at the one it came from. It returns the old position
    by value now, as post-increment is defined to. chain_iterator's
    `operator smart_ptr_t &` reinterpreted its IntrusivePtr<Node const>
    base as an IntrusivePtr<Node>; the three assignments that used it
    go through the base the object actually has, and nothing else in the
    tree wanted it.

  - Scale::snap2Scale() reads pitchScaleDeltas[ 0 ] through a
    min_element() over [ 0, totalTones ), which is empty when a scale
    has no tones. Its one caller returns before it in that case, from
    another translation unit -- so the precondition is stated here now,
    which answers -Wmaybe-uninitialized and gives a checked build an
    assert if a second caller forgets.

  - One using-declaration named for an assertion that only exists under
    _DEBUG, which is no build off MSVC.

Verified from scratch on aarch64 Linux, GCC 15.2: Debug and Release both
build with no diagnostics at all, and ctest is 260/264 in Debug. The four
that fail are offscreen GUI renders that segfault the same way on
d1cd5c4, before any of this.

+++ src/core/automatedModuleChain.cpp
@@ -23,9 +23,10 @@ ModuleChainParameter
 AutomatedModuleChain::getParameterForIndex(std::uint8_t const moduleIndex) const
 {
     auto const pModule(ModuleChainImpl::module(moduleIndex));
-    return ModuleChainParameter(!isEnd(pModule)
-                                    ? Engine::actualModule<Module const>(*pModule).effectTypeIndex()
-                                    : noModule);
+    return ModuleChainParameter(
+        !isEnd(pModule) ? static_cast<std::int8_t>(
+                              Engine::actualModule<Module const>(*pModule).effectTypeIndex())
+                        : noModule);
 }
 
 AutomatedModuleChain::ModulePtr AutomatedModuleChain::module(std::uint8_t const index)
